Fix AppImage build with Conan deps and benchmark rating type mismatch

+++ tests/benchmarks/backtracking_rate_benchmark.cpp
@@ -67,7 +67,7 @@ struct StallSnapshot {
 };
 
 struct PuzzleAnalysis {
-    int rating{0};
+    double rating{0.0};
     int clue_count{0};
     bool requires_backtracking{false};
     std::set<SolvingTechnique> techniques_used;
